% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{accept_offer}
\alias{accept_offer}
\title{Accept an offer}
\usage{
accept_offer(connection, offer_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{offer_id}{The id of the offer}
}
\value{
A FGResponse object
}
\description{
Accept an offer
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
accept_offer(conn, "of_123")
}
