% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\name{ncol,RasterArray-method}
\alias{ncol,RasterArray-method}
\alias{nrow,RasterArray-method}
\title{Number of columns and rows of a RasterArray}
\usage{
\S4method{ncol}{RasterArray}(x)

\S4method{nrow}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterLayer} class object.}
}
\value{
A \code{numeric} value of the number of columns and rows.
}
\description{
Unlike the \code{ncol} and \code{nrow} functions of the raster package (\code{\link[raster]{ncell}}), this function returns the number of columns and rows of the \code{RasterArray} container, rather than the dimensions of the contained \code{RasterLayer}s.
}
\examples{
data(clim)
ncol(clim)
nrow(clim)
}
