% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{aggregate,RasterArray-method}
\alias{aggregate,RasterArray-method}
\title{Aggregate raster cells in a RasterArray object}
\usage{
\S4method{aggregate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{...}{arguments passed to the \code{\link[raster]{aggregate}} function.}
}
\value{
An aggregated \code{RasterArray} class object.
}
\description{
The method is inherited from the \code{RasterStack} class.
}
\examples{
data(dems)
agg <- aggregate(dems, 5)

}
