% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,RasterArray-method}
\title{Aggregate raster cells in a \code{\link[chronosphere:RasterArray-class]{RasterArray}} object}
\usage{
\S4method{aggregate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[chronosphere:RasterArray-class]{RasterArray}}-class object.}

\item{...}{arguments passed to the \code{\link[raster]{aggregate}} function.}
}
\value{
An aggregated \code{\link[chronosphere:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the \code{\link[raster:raster]{RasterStack}} class.
}
\examples{
data(dems)
agg <- aggregate(dems, 5)
}
