% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\docType{methods}
\name{apply}
\alias{apply}
\alias{apply,RasterArray-method}
\title{Apply-type iterator for RasterArrays}
\usage{
apply(X, MARGIN, FUN, ...)

\S4method{apply}{RasterArray}(X, MARGIN, FUN, ...)
}
\description{
The function implements the apply-type iterators for the RasterArray class. Output values are constrained to RasterArrays, whenever possible. 
Not yet implemented for multidimensional MARGINs.
}
\examples{
# not yet
a<- cbind(demo, demo)
}
