% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{length,RasterArray-method}
\alias{length,RasterArray-method}
\alias{nlayers,RasterArray-method}
\title{Number of elements or layers in a RasterArray class object}
\usage{
\S4method{length}{RasterArray}(x)

\S4method{nlayers}{RasterArray}(x)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}
}
\description{
Function to return the length of the array in which \code{RasterLayers} are organized.
}
\details{
The length() function returns the number elements that should be present based on the array structure itself, and not the total number of values stored in the object (such as the \code{length} method of \code{RasterStack}s). As the object can contain missing values, the number of actual layers can be queried with \code{\link{nlayers}}.
}
\examples{
data(demo)
# omit third element
demo[3] <- NA
# number of elements in the RasterArray
length(demo)
# remaining number values in the stack 
length(demo@stack)
# the number of remaining layers in the RasterArray
nlayers(demo)

}
