% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-cast.R, R/SpatialArray-cast.R
\name{as.data.frame.RasterArray}
\alias{as.data.frame.RasterArray}
\alias{as.data.frame.SpatialArray}
\title{S3-type method for \code{\link{RasterArray}} and \code{\link{Spatialarray}}}
\usage{
\method{as.data.frame}{RasterArray}(x, row.names = NULL, optional = FALSE, ...)

\method{as.data.frame}{SpatialArray}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{RasterArray}} and \code{\link{Spatialarray}} class object.}

\item{row.names}{Argument to define the rownames of the resulting \code{data.frame}.}

\item{optional}{\code{logical} Flag to reset the rownames and colnaems attributes.}

\item{...}{additional arguments passed to and from methods.}
}
\value{
A \code{data.frame} class object.
}
\description{
Convert RasterArray class objects to data.frames allowing View(), head() and tail() to work.
}
\details{
Formal conversion method transforming the proxy object to a data.frame.
}
\examples{
data(dems)
df <- as.data.frame(dems)

}
