% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-cast.R
\docType{methods}
\name{as.list,RasterArray-method}
\alias{as.list,RasterArray-method}
\title{Coerce a RasterArray class object to a list}
\usage{
\S4method{as.list}{RasterArray}(x, ...)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}

\item{...}{arguments passed to or from methods.}
}
\description{
The function will return the items of the RasterArray as a list (conserving the names of the elements in the stack).
}
