% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialArray-cast.R
\name{as.list,SpatialArray-method}
\alias{as.list,SpatialArray-method}
\title{Coerce a \code{\link{SpatialArray}} class object to a list}
\usage{
\S4method{as.list}{SpatialArray}(x, ...)
}
\arguments{
\item{x}{A \\code{\link{SpatialArray}} class object.}

\item{...}{arguments passed to or from methods.}
}
\value{
A \code{list} of \code{RasterLayers}.
}
\description{
The function will return the items of the \code{\link{SpatialArray}} as a list (conserving the names of the elements in the stack).
}
