% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection.R
\name{assert_token_is_not_expired}
\alias{assert_token_is_not_expired}
\title{Helper Function to check if a FGConnection is still valid. Normally called automatically.}
\usage{
assert_token_is_not_expired(connection)
}
\arguments{
\item{connection}{The connection to be tested}
}
\value{
None
}
\description{
Helper Function to check if a FGConnection is still valid. Normally called automatically.
}
\examples{
None
}
