% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialArray-cast.R
\name{as.SpatialArray}
\alias{as.SpatialArray}
\title{Convert Spatial* objects to SpatialArrays}
\usage{
as.SpatialArray(from)
}
\arguments{
\item{from}{Object to be converted.}
}
\value{
A \code{aSpatialArray} class object.
}
\description{
The function converts \code{Spatial*}objects to \code{\link{SpatialArray}}-class objects.
}
\examples{
# data(dems)
# recent <- dems[1]
# # convert RasterLayer to RasterArray
# ra <- as.RasterArray(recent)
}
