% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_climate_change.R
\name{avg_var}
\alias{avg_var}
\title{Calculates the spatial average of a variable}
\usage{
avg_var(x, period = NULL)
}
\arguments{
\item{x}{is a raster object of the variable of interest, with each layer representing time}

\item{period}{corresponds to the name of the raster layers of x of interest. Leave blank to calculate average of all the layers.}
}
\value{
RasterLayer
}
\description{
\code{avg_var} calculates the spatial the spatial average of a variable of interest over a period of time.
}
