% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-combine.R
\name{c.RasterLayer}
\alias{c.RasterLayer}
\alias{c.RasterArray}
\title{Combine RasterLayers and one-dimensional RasterArrays}
\usage{
\method{c}{RasterLayer}(...)

\method{c}{RasterArray}(...)
}
\description{
Methods sequences that start with an NA do not yet work.
}
