% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{methods}
\name{cellStats,RasterArray-method}
\alias{cellStats,RasterArray-method}
\title{Statistics across cells in a RasterArray object}
\usage{
\S4method{cellStats}{RasterArray}(x, stat, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{stat}{A function to be applied.}

\item{...}{arguments passed to the \code{\link[raster]{cellStats}} function.}
}
\description{
The method is inherited from the \code{RasterStack} class. Postitions of layers are conserved in the output. (including missing layers)
}
\examples{
data(clim)
cellStats(clim, stat=mean, na.rm=TRUE)
}
