% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{chronosphere}
\alias{chronosphere}
\title{Evolving Earth System Variables
The purpose of the 'chronosphere' project is to increase reproducibility and establish version tracking of results from(paleo)environmental/ecological research. The inherent database maps data files and import code to programming classes and allows access to these items via files deposited in public repositories.#' 
This is still the Release Candidate version. As is R, this is free software and comes with ABSOLUTELY NO WARRANTY.  Nevertheless, notes about found bugs and suggestions are more than welcome. The chronosphere links to data and code deposited on public servers, which can be inspected using the code and data URLs of the items (run \code{datasets(master=TRUE)}). By using this package you agree that you take responsibility for the items that you download. Never run \code{fetch()} as a superuser or with administrative privileges, which can open up a security hole to be exploited in case the remote servers get hijacked.}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
