% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clim}
\alias{clim}
\title{Temperature dataset from modelled simulations}
\format{A raster brick of 3 layers corresponding to time intervals.
\describe{
  \item{X0}{temperature data for 0 Ma}
  \item{X4}{temperature data for 4 Ma}
  \item{X11}{temperature data for 11 Ma}
}}
\source{
Valdes simulations
}
\usage{
clim
}
\description{
A dataset containing temperature data from Paul Valdes and Chris Scotese for the time
0, 4 and 11 Ma.
}
\keyword{datasets}
