% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coasts}
\alias{coasts}
\title{PaleoMAP PaleoCoastlines (demo)
This dataset is important.}
\format{
An object of class \code{SpatialArray} with 5 rows and 2 columns.
}
\usage{
data(coasts)
}
\description{
PaleoMAP PaleoCoastlines (demo)
This dataset is important.
}
\keyword{datasets}
