% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{colnames,RasterArray-method}
\alias{colnames,RasterArray-method}
\alias{colnames<-,RasterArray-method}
\title{Column names of two-dimensional RasterArray or SpArray}
\usage{
\S4method{colnames}{RasterArray}(x)

\S4method{colnames}{RasterArray}(x) <- value
}
\arguments{
\item{x}{\code{RasterArray} or \code{SpArray} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of column names or \code{NULL}.
}
\description{
Get or set the column names of two-dimensional RasterArray or SpArray objects
}
\examples{
data(clim)
colnames(clim)
colnames(clim) <- c("a", "b")
}
