% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-combine.R, R/RasterArray-combine.R,
%   R/SpatialArray-combine.R
\name{combine}
\alias{combine}
\alias{combine,XArray-method}
\alias{combine,RasterLayer-method}
\alias{combine,VectorSpatialClasses-method}
\title{Combine RasterLayers and one-dimensional RasterArrays and Spatial* objects with one-dimensinoal SpatialArrays}
\usage{
combine(x, ...)

\S4method{combine}{XArray}(x, ...)

\S4method{combine}{RasterLayer}(x, ...)

\S4method{combine}{VectorSpatialClasses}(x, ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterArray} objects or Spatial* and \code{\link{SpatialArray}} objects to combine.}

\item{...}{additional objects to combine.}
}
\value{
A \code{\link{RasterArray}} or \code{\link{SpatialArray}}class object.
}
\description{
Methods sequences that start with an NA do not yet work.
}
\examples{
data(dems)
a <- combine(dems[1], dems[2])
}
