% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect_with_pat}
\alias{connect_with_pat}
\title{Get a FastGenomics connection object to a specific FASTGenomics instance.
This method is less secure than using connect.
Use this method only if save_personal_access_token does not work, e.g. because your keyring does not work or you work in a Jupyter Notebook.}
\usage{
connect_with_pat(base_url, email, pat)
}
\arguments{
\item{base_url}{The url of the instance, e.g. https://fastgenomics.org/}

\item{email}{The email address of your account}

\item{pat}{Your PAT, NEVER share this or store it in your history etc.}
}
\value{
a connection object
}
\description{
WARNINIG: NEVER commit passwords or Tokens to a GIT repository or share them in any way! The token can be used to do any action on your behalf.
In the background, your PAT is used.
}
\examples{
fastgenomicsRclient::save_personal_access_token("https://fastgenomics.org/", "user@example.com")
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "user@example.com")
}
