% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{create_app}
\alias{create_app}
\title{Creates a new app}
\usage{
create_app(connection, source_image_name, image_name = "",
  registry = "", username = "", password = "")
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{source_image_name}{The name of the app image, must inlcude a tag. E.g. yourapp:latest. Must not contain the name of the registry.}

\item{image_name}{The name to be used in FASTGenomics, e.g. yourapp:v1. Note that your username is always prepended, e.g. someuser/yourapp:v1.}

\item{registry}{The registry to be used, leave blank if you use the docker hub. Do not prepend http(s), egg someregistry.example.com}

\item{username}{The username of the registry, leave blank if no auth is required. FG uses this information ONLY to pull the image, neither username nor password are stored permanently.}

\item{password}{The password of the registry, leave blank if no auth is required.}
}
\value{
An app or a validation error
}
\description{
Pulls an app from a docker registry (e.g. the docker hub or a private registry) and installs it in FASTGenomics. The image will be pulled by FASTGenomics
After submitting your app, the app will be validated on the server, this can take a while. The app is usable when the validation has been completed
Use poll_app_until_validated to wait for the validation to complete.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
result <- create_app(default_conn, "library/busybox:latest", registry = "registry.hub.docker.com" , image_name = "my_image_name:1", username = "a", password = "b" )
status <- poll_app_until_validated(connection, result) # will return FALSE, since neither image nor credentials are valid
}
