% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-sparse.R
\name{create_dataset_df}
\alias{create_dataset_df}
\title{Submits a data set from a sparse matrix of dgTmatrix type and
gene/cell metadata.  dgTMatrix has to have named dimensions, cell
and gene metadata must have cellId and geneId columns respectively.
The names of the dimensions in the matrix must be a subset of the
respective gene/cell metadata data frames.}
\usage{
create_dataset_df(connection, matrix, cell_metadata, gene_metadata,
  gene_nomenclature, organism_id, title, zipfiles = TRUE,
  description = "", short_description = "", batch_column = "",
  optional_parameters = FGDatasetUploadParameters(), tmpdir = NULL)
}
\arguments{
\item{connection}{FASTGenomics connection object}

\item{matrix}{sparseMatrix storing the expression table.  The first
dimension is assumed to be genes and the second must be cell
names.}

\item{cell_metadata}{dataframe with cell metadata, must have a
cellId column.}

\item{gene_metadata}{dataframe with gene metadata, must have a
geneId column.}

\item{organism_id}{One of 9606 (Mouse) and 10090 (Human)}

\item{title}{The title of the data set}

\item{zipfiles}{Weather to compress files before submitting,
Default: TRUE}

\item{description}{long description, Default: ""}

\item{short_description}{short description, Default: ""}

\item{batch_column}{The column in the gene_metadata which is used to hold batch information. Can be "" for no batch information}

\item{optional_parameters}{Object representing further optional
parameters, see \code{\link{FGDatasetUploadParameters}},
Default: NULL
The properties batch_column, gene_metadata and cell_metadata of this object are always ignored}

\item{tmpdir}{The location of temporary files, Default: NULL}
}
\value{
either of \code{\link{FGResponse}},
    \code{\link{FGErrorResponse}},
    \code{\link{FGErrorModelResponse}},
    \code{\link{FGValidationProblem}}.
}
\description{
Submits a data set from a sparse matrix of dgTmatrix type and
gene/cell metadata.  dgTMatrix has to have named dimensions, cell
and gene metadata must have cellId and geneId columns respectively.
The names of the dimensions in the matrix must be a subset of the
respective gene/cell metadata data frames.
}
