% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-sparse.R
\name{create_dataset_from_seurat}
\alias{create_dataset_from_seurat}
\title{Uploads a seurat dataset.  Runs \code{\link{create_dataset_df}}
with matrix, cell_metadata and gene_metadata inferred from the
seurat object.}
\usage{
create_dataset_from_seurat(connection, seurat_obj, ...)
}
\arguments{
\item{connection}{FASTGenomics connection object}

\item{seurat_obj}{Seurat object to submitted}

\item{...}{Other parameters passed to \code{\link{create_dataset_df}}}
}
\description{
Uploads a seurat dataset.  Runs \code{\link{create_dataset_df}}
with matrix, cell_metadata and gene_metadata inferred from the
seurat object.
}
