% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{create_share}
\alias{create_share}
\title{Create a new share}
\usage{
create_share(connection, dataset_id, message, license, send_to)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{dataset_id}{The id of the dataset to be shared. You must own this dataset. Should start with ds_}

\item{message}{The message to be presented to the recipients}

\item{license}{The license to be presented to the recipients}

\item{send_to}{List of email addresses}
}
\value{
A FGResponse object
}
\description{
Create a new share
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
share <- create_share(conn, "dts_0f91c4f034e1471fa25f1078d23e1a97", message = "no msg", license = "license", send_to = list("a2@example.com"))
print(share@content) # the new share
}
