% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{create_workflow}
\alias{create_workflow}
\title{Creates a new workflow on the server.}
\usage{
create_workflow(connection, path_to_workflow)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{path_to_workflow}{The file path to the edit model JSON.}
}
\value{
A FGResponse object if no errors, otherwisea a FGErrorResponse
}
\description{
This method expects a file with valid json as input, since creating a workflow is best done by hand. Use a good editor like VS Code for this task.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
 # get a workflow as example
workflow <- fastgenomicsRclient::get_edit_model_of_workflow(connection, "wf_abc")
print(workflows@content) # the workflows as an easily usable model
path <- "workflow.json"
save_workflow_as_file(workflow, path) # save the workflow on disk
# make edits on the file...
result <- create_workflow(connection, path) # create the workflow
# check that the Result is not an FGErrorResponse, if it is an error check the "errors" field on the object.
}
