% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{methods}
\name{crop,RasterArray-method}
\alias{crop,RasterArray-method}
\title{Crop a RasterArray object}
\usage{
\S4method{crop}{RasterArray}(x, y, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{y}{an xtent object, or any object from which an Extent object can be extracted (see Details)}

\item{...}{arguments passed to the \code{\link[raster]{crop}} function.}
}
\description{
The method is inherited from the \code{RasterStack} class.
}
