% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{dataindex}
\alias{dataindex}
\title{Deprecated function for data fetching}
\usage{
dataindex(...)
}
\arguments{
\item{...}{Arguments passed to the \code{\link{datasets}} function.}
}
\value{
The output of the \code{\link{datasets}} function.
}
\description{
This function will be removed in version 0.3.0.
}
