% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{decline_offer}
\alias{decline_offer}
\title{Decline an offer}
\usage{
decline_offer(connection, offer_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{offer_id}{The id of the offer}
}
\value{
A FGResponse object
}
\description{
Decline an offer
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
decline_offer(conn, "of_123")
}
