% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{delete_share}
\alias{delete_share}
\title{Delete a share}
\usage{
delete_share(connection, share_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{share_id}{The share to be deleted.}
}
\value{
A FGResponse object
}
\description{
Note that all recipients must decline or have their access revoked before the share can be deleted.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
share <- create_share(conn, "dts_0f91c4f034e1471fa25f1078d23e1a97", message = "no msg", license = "license", send_to = list("a2@example.com"))
delete_share(conn, share@Id)
}
