% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{delete_workflow}
\alias{delete_workflow}
\title{Delete a specific workflow.}
\usage{
delete_workflow(connection, workflow_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{workflow_id}{The id of the workflow, usually looks like wf_.....}
}
\value{
A FGResponse object
}
\description{
Delete a specific workflow.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
workflows <- fastgenomicsRclient::delete_workflow(connection, "wf_abc")
}
