% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo}
\alias{demo}
\title{PaleoDEM rasters from the reconstructions Chris Scotese.}
\format{A RasterArray with 10 layers.}
\source{
Chris Scotese
}
\usage{
demo
}
\description{
A dataset containing the paleoDEM reconstructions of Chris Scotese for the time interval 0 - 45 Ma.
}
\keyword{datasets}
