% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dems}
\alias{dems}
\title{PaleoDEM rasters from the reconstructions Chris Scotese.}
\format{
A \code{\link[chronosphere:RasterArray-class]{RasterArray}} with 10 layers.
}
\source{
Scotese, C. R. Wright, N. (2018). PALEOMAP Paleodigital Elevation Models (PaleoDEMS) for the Phanerozoic. URL: \url{https://www.earthbyte.org/paleodem-resource-scotese-and-wright-2018/}
}
\usage{
data(dems)
}
\description{
A dataset containing the paleoDEM reconstructions of Chris Scotese for the time interval 0 - 45 Ma.
}
\details{
This is a subset of the total dataset, use \code{fetch(dat="paleomap", var="dem", res=1, ver="20190719")} to get the whole set.
}
\keyword{datasets}
