% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-xattrib.R
\name{dim,XArray-method}
\alias{dim,XArray-method}
\title{Dimensions of \code{\link{SpatialArray}} or \code{\link{RasterArray}} objects}
\usage{
\S4method{dim}{XArray}(x)
}
\arguments{
\item{x}{A \code{\link{SpatialArray}} or \code{\link{RasterArray}} class object.}
}
\value{
A \code{numeric} vector.
}
\description{
The function returns the dimensions of the array in which \code{RasterLayer}s or \code{Spatial*} objects are organized.
}
\examples{
data(dems)
dim(dems)
data(coasts)
dim(coasts)
}
