% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{dimlayer}
\alias{dimlayer}
\alias{dimlayer,RasterArray-method}
\title{Dimensions of RasterLayers in a RasterArray object}
\usage{
dimlayer(x, ...)

\S4method{dimlayer}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterLayer} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\description{
The funcion will return the dimensions RasterLayers
}
