% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray.R
\docType{methods}
\name{dimnames,RasterArray-method}
\alias{dimnames,RasterArray-method}
\alias{dimnames<-,RasterArray,ANY-method}
\title{Names of multidimensional RasterArray or SpArray objects.}
\usage{
\S4method{dimnames}{RasterArray}(x)

\S4method{dimnames}{RasterArray,ANY}(x) <- value
}
\arguments{
\item{x}{\code{RasterArray} or \code{SpArray} object.}

\item{value}{\code{character} vector.}
}
\description{
Get or set the dimnames of multidimensional RasterArray or SpArray objects
}
\examples{
# an example
}
