% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{data}
\name{disaggregate}
\alias{disaggregate}
\alias{disaggregate,RasterArray-method}
\title{Disaggregate raster cells in a RasterArray object}
\format{An object of class \code{standardGeneric} of length 1.}
\usage{
disaggregate

\S4method{disaggregate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{...}{arguments passed to the \code{\link[raster]{disaggregate}} function.}
}
\value{
A disaggregated \code{RasterArray} class object.
}
\description{
The method is inherited from the \code{RasterStack} class.
}
\examples{
data(dems)
disagg <- disaggregate(dems, 3)
}
\keyword{datasets}
