% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{extent}
\alias{extent}
\alias{extent,RasterArray-method}
\title{Extent of a \code{\link[chronosphere:RasterArray-class]{RasterArray}} object}
\usage{
\S4method{extent}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[chronosphere:RasterArray-class]{RasterArray}}-class object.}

\item{...}{arguments passed to the \code{\link[raster]{extent}} function.}
}
\value{
An aggregated \code{\link[chronosphere:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the \code{\link[raster:raster]{RasterStack}} class.
}
\examples{
data(dems)
agg <- extent(dems)
}
