% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{methods}
\name{minValue,RasterArray-method}
\alias{minValue,RasterArray-method}
\alias{maxValue,RasterArray-method}
\title{Minimum and maximum values in a RasterArray object}
\usage{
\S4method{minValue}{RasterArray}(x, vec = FALSE)

\S4method{maxValue}{RasterArray}(x, vec = FALSE)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{vec}{Should the dimensions of the \code{RasterArray} be omitted?}

\item{...}{arguments passed to the \code{\link[raster]{cellStats}} function.}
}
\description{
The method is inherited from the \code{RasterStack} class. Postitions of layers are conserved in the output. (including missing layers)
}
\examples{
data(dems)
rangeVals <- cbind(
  minValue(dems),
  maxValue(dems)
)
}
