% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection.R
\docType{class}
\name{FGConnection-class}
\alias{FGConnection-class}
\title{A FGConnection}
\value{
class FGConnection
}
\description{
Make sure that you never share an environment which contains such an object.
}
\section{Slots}{

\describe{
\item{\code{base_url}}{the URL.}

\item{\code{bearer_token}}{the token.}
}}

\examples{
None
}
