% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_connection.R
\docType{data}
\name{FGConnection}
\alias{FGConnection}
\title{A FGConnection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FGConnection
}
\value{
class FGConnection
}
\description{
Make sure that you never share an environment which contains such an object.
}
\section{Slots}{

\describe{
\item{\code{base_url}}{the URL.}

\item{\code{pat}}{the personal access token.}

\item{\code{the}}{email address of the account}
}}

\examples{
None
}
\keyword{datasets}
