% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGResponse.R
\docType{class}
\name{FGErrorModelResponse-class}
\alias{FGErrorModelResponse-class}
\title{Validation error}
\value{
class FGErrorModelResponse
}
\description{
Validation error
}
\section{Slots}{

\describe{
\item{\code{content}}{the response}

\item{\code{path}}{the url}

\item{\code{validation_errors}}{a list of validation errors. Fixe these errors and try your operation again.}
}}

\examples{
None
}
