% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGErrorResponse.R
\docType{class}
\name{FGErrorResponse-class}
\alias{FGErrorResponse-class}
\title{Information about FG Errors}
\value{
class FGErrorResponse
}
\description{
Information about FG Errors
}
\section{Slots}{

\describe{
\item{\code{content}}{the response}

\item{\code{path}}{the url}

\item{\code{validation_errors}}{a list of validation errors. Fixe these errors and try your operation again.}
}}

\examples{
None
}
