% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGResponse.R
\docType{class}
\name{FGResponse-class}
\alias{FGResponse-class}
\title{A FG response after a successful request}
\value{
class FGResponse
}
\description{
A FG response after a successful request
}
\section{Slots}{

\describe{
\item{\code{content}}{The data}

\item{\code{path}}{The url used to get the data}

\item{\code{response}}{the raw response.}

\item{\code{DataType}}{The type of data}

\item{\code{Id}}{The id of the object, if the content is a list, this is blank}
}}

\examples{
None
}
