% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGValidationProblem.R
\docType{class}
\name{FGValidationProblem-class}
\alias{FGValidationProblem-class}
\title{Encapsulates RFC 7807 Problem details, with the errors field added}
\value{
FGValidationProblem
}
\description{
See https://tools.ietf.org/html/rfc7807 for details
}
\section{Slots}{

\describe{
\item{\code{errors}}{List of errors, human readable}

\item{\code{title}}{A short, human-readable summary of the problem type.}

\item{\code{type}}{the problem type}

\item{\code{status}}{The HTTP status code}

\item{\code{detail}}{human-readable explanation specific to this occurrence of the problem.}

\item{\code{instance}}{A URI reference that identifies the specific occurrence of the problem.}
}}

\examples{
None
}
