% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_bib}
\alias{generate_bib}
\title{Generate .bib file from \code{data.frame}}
\usage{
generate_bib(
  x,
  output_path = ".",
  type = "publication_type",
  combine = TRUE,
  return_fnames = TRUE
)
}
\arguments{
\item{x}{(\code{data.frame}) references with specific headers. See details for more information.}

\item{output_path}{folder where to save generated .bib file.}

\item{type}{(\code{character}) type column in \code{x} which specifies the publication type.}

\item{combine}{(\code{logical}) Should the references be save on one file or separated in several ones. See details for more information.}

\item{return_fnames}{(\code{logical}) Should the files names of the generated bib files be returned?}
}
\value{

}
\description{
Generate .bib file from \code{data.frame}
}
\details{
The \code{generate_bib} function allows conversion of data.frames to BibLaTex entries. Our specifications
are based on the tempalte used by the Paleobiology Database. The following entry types are currently supported, with the specified column names:

\itemize{
\item{journal article}{ Columns: \code{reftitle}, \code{author}, \code{pubtitle}, \code{pubvol}, \code{page}, \code{pubyear}, \code{doi}}
\item{book}{ Columns: \code{pubtitle}, \code{author}, \code{pubyr}, \code{publisher}, \code{pubcity}}
\item{book chapter}{ Columns: \code{reftitle}, \code{pubtitle}, \code{author}, \code{editors}, \code{pubyr}, \code{publisher}, \code{pubcity}, \code{page}}
\item{book series}{ Columns: \code{reftitle} for the book title, \code{pubtitle} for the series title, \code{author}, \code{pubno}, \code{pubyr}, \code{publisher}, \code{pubcity}}
\item{guidebook}{ See \code{book}}
\item{thesis}{ Columns: \code{reftitle}, \code{author}, \code{pubyr}, \code{pubtitle} for the institution}
\item{misc}{ Columns: \code{reftitle}, \code{author}, \code{pubyr}}
\item{abstract}{ See \code{misc}}
\item{news article}{ See \code{misc}}
\item{compendium}{S ee \code{misc}}
\item{unpublished}{ See \code{misc}}
}

Note: 
\itemize{
\item{\code{author} can also be provided as \code{author1init}, \code{author1last}, \code{author2init}, \code{author2last}, \code{otherauthors}}
\item{\code{page} can also be provided as \code{firstpage} and \code{lastpage}}
}
}
\examples{

# refs <- read.csv("https://paleobiodb.org/data1.2/occs/refs.csv?base_name=Scleractinia&occs_created_after=2020-01-01&select=occs&all_records",
# # encoding required to allow reading of special characeters, e.g. accents characters such as é, è
# encoding="UTF-8")[1:2,]
# 
# generate_bib(refs)

}
