% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(
  metadata = list(title = "Sample title", authors = c("Jane Doe", "John Doe"),
    corresauth = 1, corresemail = "jane.doe@fau.de", affiliation =
    c("Friedrich Alexander Universität", "University of Birmingham"), abstract =
    "This is where my abstract goes...", keywords = "data, paleobiology, fossil"),
  specs = list(source = "The Paleobiology Database", url =
    "https://paleobiodb.org/data1.2/occs/list.csv?datainfo&rowcount&interval=\\n                            Ediacaran,Holocene&show=class,classext,genus,subgenus,coll,coords,loc,paleoloc,\\n                            strat,stratext,lith,env,ref,crmod,timebins,timecompare",
    download_date = "2020-06-04", dataformat = "Processed", dataaccess =
    "Available with this report", taxa = "Multiple", location = "Global", time =
    "Cambrian to Recent"),
  data_refs = refs,
  enterer_names,
  output_path = file.path("output"),
  output_file = "test.pdf",
  create_dir = TRUE,
  template = NULL,
  in_header = NULL,
  verbose = FALSE
)
}
\description{
Generate report
}
