% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_app}
\alias{get_app}
\title{Get an app}
\usage{
get_app(connection, app_id)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{app_id}{the id of the app or a FGResponse object}
}
\value{
class FGResponse
}
\description{
Get an app
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
app <- fastgenomicsRclient::get_apps(connection, "abc")
print(app@content) # the app
}
