% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_apps}
\alias{get_apps}
\title{Get all apps}
\usage{
get_apps(connection, scope = "All")
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{scope}{Filters the apps by their scope. Possible Values are: 'All': return all apps, 'Private': Only your personal apps, 'Public': Only public apps}
}
\value{
A FGResponse object
}
\description{
This list does not contain rejected apps, only valid, usable apps
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
apps <- fastgenomicsRclient::get_app(connection)
print(apps@content) # all apps available to you
}
