% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a specific dataset}
\usage{
get_dataset(connection, dataset_id)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{dataset_id}{The id of the dataset, usually starting with dts_*****}
}
\value{
A FGResponse object
}
\description{
Gets information about a specific dataset. This can also be used to obtain information about a failed dataset upload
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
datasets <- fastgenomicsRclient::get_dataset(connection, "dts_abc")
print(datasets@content) # the dataset
}
