% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Get all datasets}
\usage{
get_datasets(connection, scope = "All")
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{scope}{Filters the datasets by their scope. Possible Values are: 'All': return all datasets, 'Private': Only your personal datasets, 'Public': Only public datasets}
}
\value{
A FGResponse object
}
\description{
This list does not contain rejected datasets, only valid, usable datasets.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
datasets <- fastgenomicsRclient::get_datasets(connection)
print(datasets@content) # all datasets available to you
}
