% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_edit_model_of_workflow}
\alias{get_edit_model_of_workflow}
\title{Get a specific workflow as an editable model}
\usage{
get_edit_model_of_workflow(connection, workflow_id)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{workflow_id}{The id of the workflow, usually looks like wf_.....}
}
\value{
A FGResponse object if no errors, otherwisea a FGErrorResponse
}
\description{
This model is suitable to be used as a base for new workflows.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
workflow <- fastgenomicsRclient::get_edit_model_of_workflow(connection, "wf_abc")
print(workflows@content) # the workflows as an easily usable model
}
