% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_known_defaults}
\alias{get_known_defaults}
\title{Get a list of all known defaults}
\usage{
get_known_defaults(connection)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}
}
\value{
a list of the known defaults
}
\description{
Get a list of all known defaults
}
\examples{
None
}
