% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{get_offer}
\alias{get_offer}
\title{Get an offer}
\usage{
get_offer(connection, offer_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{offer_id}{The id of the offer}
}
\value{
A FGResponse object
}
\description{
Get an offer
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
get_offer(conn, "offer")
}
