% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{get_offers}
\alias{get_offers}
\title{Get up to 100 offers from the client}
\usage{
get_offers(connection, take = 100, skip = 0)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{take}{How many objects to take}

\item{skip}{How many objects to skip}
}
\value{
A FGResponse object
}
\description{
Get up to 100 offers from the client
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
offers <- fastgenomicsRclient::get_offers(connection)
print(offers@content) # all offers available to you
}
