% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{get_share}
\alias{get_share}
\title{Get a share}
\usage{
get_share(connection, share_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{share_id}{The id of the share}
}
\value{
A FGResponse object
}
\description{
Get a share
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
share <- create_share(conn, "dts_0f91c4f034e1471fa25f1078d23e1a97", message = "no msg", license = "license", send_to = list("a2@example.com"))
get_share(conn, share@Id)
}
