% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{get_shares}
\alias{get_shares}
\title{Get all shares}
\usage{
get_shares(connection)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}
}
\value{
A FGResponse object
}
\description{
Pagination is handled in the client
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
shares <- fastgenomicsRclient::get_shares(connection)
print(shares@content) # all shares available to you
}
