% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_valid_current_normalization_status}
\alias{get_valid_current_normalization_status}
\title{Get a list of all supported normalization schemes}
\usage{
get_valid_current_normalization_status(connection)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}
}
\value{
a list of supported normalization schemes
}
\description{
Get a list of all supported normalization schemes
}
\examples{
None
}
