% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_valid_gene_nomenclatures}
\alias{get_valid_gene_nomenclatures}
\title{Get a list of all supported gene nomenclatures}
\usage{
get_valid_gene_nomenclatures(connection)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}
}
\value{
a list of the valid gene nomenclatures
}
\description{
Get a list of all supported gene nomenclatures
}
\examples{
None
}
