% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_valid_matrix_formats}
\alias{get_valid_matrix_formats}
\title{Get a list of all supported matrix formats}
\usage{
get_valid_matrix_formats(connection)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}
}
\value{
a list of the valid matrix formats
}
\description{
Get a list of all supported matrix formats
}
\examples{
None
}
