% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_valid_technologies}
\alias{get_valid_technologies}
\title{Get a list of all supported technologies}
\usage{
get_valid_technologies(connection)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}
}
\value{
a list of the valid technologies
}
\description{
Get a list of all supported technologies
}
\examples{
None
}
