% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_workflow}
\alias{get_workflow}
\title{Get a specific workflow}
\usage{
get_workflow(connection, workflow_id, fullDetail = FALSE)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{workflow_id}{The id of the workflow, usually looks like wf_.....}

\item{fullDetail}{If TRUE, the calculationflow, screenflow and parameters are included.}
}
\value{
A FGResponse object
}
\description{
This model is not suitable for editing OR as base template for new workflows, call get_edit_model_of_workflow to obtain such a workflow!
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
workflows <- fastgenomicsRclient::get_workflow(connection, "wf_abc")
print(workflows@content) # the workflow
}
