% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_workflows}
\alias{get_workflows}
\title{Get all workflows}
\usage{
get_workflows(connection, scope = "All")
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{scope}{Filters the workflows by their scope. Possible Values are: 'All': return all workflows, 'Private': Only your personal workflows, 'Public': Only public workflows}
}
\value{
A FGResponse object
}
\description{
This list does only contain valid, usable datasets
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
workflows <- fastgenomicsRclient::get_workflows(connection)
print(workflows@content) # all workflows available to you
}
