% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{gplates_reconstruct_coastlines}
\alias{gplates_reconstruct_coastlines}
\title{Reconstruct coastlines
Retrieve reconstructed coastline polygons for defined ages}
\usage{
gplates_reconstruct_coastlines(age, model = "PALEOMAP", verbose = TRUE)
}
\arguments{
\item{age}{is the age in Ma at which the points will be reconstructed}

\item{model}{is the reconstruction model. The default is "PALEOMAP". Add more details about additional models here}

\item{verbose}{Should the function output urls?}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Reconstruct coastlines
Retrieve reconstructed coastline polygons for defined ages
}
\examples{
gplates_reconstruct_coastlines(140)
}
