% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{gplates_reconstruct_points}
\alias{gplates_reconstruct_points}
\title{Reconstruct points}
\usage{
gplates_reconstruct_points(coords, age, model = "PALEOMAP",
  reverse = FALSE, verbose = TRUE)
}
\arguments{
\item{coords}{are the coordinates to be reconstructed. Can be a vector with longitude and latitude representing
a single point or a matrix/dataframe with the first column as longitude and second column as latitude}

\item{age}{is the age in Ma at which the points will be reconstructed}

\item{model}{is the reconstruction model. The default is "PALEOMAP". Add more details about additional models here}

\item{reverse}{the flag to control the direction of reconstruction. If reverse = TRUE, the function will 
calculate the present-day coordinates of the given paleo-coordinates.}

\item{verbose}{Should the function output urls?}
}
\value{
matrix with longitude and latitude
}
\description{
Reconstruct the geographic locations from present day coordinates back to their paleo-positions. 
Each location will be assigned a plate id and moved back in time using the chosen reconstruction model.
}
\details{
Adapted from GPlates Web Service (need to find out how to reference them)
}
\examples{
gplates_reconstruct_points(c(95, 54), 140)

xy <-cbind(long=c(95,142), lat=c(54, -33))
gplates_reconstruct_points(xy, 140)
}
