% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{gplates_reconstruct_polygon}
\alias{gplates_reconstruct_polygon}
\title{reconstructing polygons}
\usage{
gplates_reconstruct_polygon(sp, age, model = "PALEOMAP",
  verbose = TRUE)
}
\arguments{
\item{sp}{is a SpatialPolygonsDataFrame}

\item{verbose}{Should the function output urls?}
}
\description{
reconstructing polygons
}
