% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{gplates_reconstruct_static_polygons}
\alias{gplates_reconstruct_static_polygons}
\title{reconstruct static polygons}
\usage{
gplates_reconstruct_static_polygons(age, model = "PALEOMAP",
  verbose = TRUE)
}
\arguments{
\item{age}{is the age in Ma at which the points will be reconstructed}

\item{model}{is the reconstruction model. The default is "PALEOMAP". Add more details about additional models here}

\item{verbose}{Should the function output urls?}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Retrieve reconstructed static polygons for defined ages
}
\examples{
gplates_reconstruct_static_polygons(140)

}
