% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\name{is.na.RasterArray}
\alias{is.na.RasterArray}
\title{Positions of missing values in a RasterArray object}
\usage{
\method{is.na}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}
}
\description{
The function behaves similar to the regular \code{is.na()} function applied to the proxy object of a \code{RasterArray}.
}
\examples{
data(demo)
demo[2] <- NA
is.na(demo)

}
