% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\name{layers}
\alias{layers}
\title{Names of RasterArray's Layers in the stack}
\usage{
layers(x, ...)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}
}
\description{
Names of RasterArray's Layers in the stack
}
