% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mapedge}
\alias{mapedge}
\title{Function to quickly draft the edge of the equirectangular projection}
\usage{
mapedge(x = 360, y = 180, xmin = -180, xmax = 180, ymin = -90, ymax = 90)
}
\arguments{
\item{x}{(\code{numeric}) Number of segments in the x (longitude) dimension.}

\item{y}{(\code{numeric}) Number of segments in the y (latitude) dimension.}

\item{xmin}{(\code{numeric}) Minimum value of x (longitude).}

\item{xmax}{(\code{numeric}) Minimum value of x (longitude).}

\item{ymin}{(\code{numeric}) Maximum value of y (latitude).}

\item{ymax}{(\code{numeric}) Maximum value of y (latitude).}
}
\value{
A \code{SpatialPolygons} class object.
}
\description{
Function to plot the edge of a map with different projections.
}
\examples{
# requires rgdal
edge <- mapedge()
molledge <- spTransform(edge, CRS("+proj=moll"))

}
