% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{modify_workflow}
\alias{modify_workflow}
\title{Modifies an existing workflow}
\usage{
modify_workflow(connection, path_to_workflow, workflow_id)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{path_to_workflow}{The file path to the edit model JSON.}

\item{workflow_id}{The id of the workflow to be updated}
}
\value{
A FGResponse object if no errors, otherwisea a FGErrorResponse
}
\description{
Your modifications will not affect any existing analyses or analysistemplates. A successfull modification will increase the version field of the workflow.
You can update an analysistemplate through AnalysisApi: /api/v1/analysistemplates/{id}/update_workflow
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
 # get a workflow as example
workflow <- fastgenomicsRclient::get_edit_model_of_workflow(connection, "wf_abc")
print(workflows@content) # the workflows as an easily usable model
path <- "workflow.json"
save_workflow_as_file(workflow, path) # save the workflow on disk
# make edits on the file...
result <- create_workflow(connection, path) # create the workflow
updated_workflow <- modify_workflow(connection, path, result@content[["id"]]) # update the workflow
print(updated_workflow@content[["version"]]) # should be 2.0
}
