% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{multiextract}
\alias{multiextract}
\title{Extraction of values from a multiple RasterLayers}
\usage{
multiextract(x, y, by, lng = "plng", lat = "plat", force = NULL)
}
\arguments{
\item{x}{(\code{matrix} or \code{data.frame}). The data table containing the coordinates and (optionally) 
the indices or names of the associated \code{RasterLayer}s in \code{y}.}

\item{y}{(\code{RasterArray}). A set of \code{RasterLayers} that are associated with the rows of \code{x}.}

\item{by}{(\code{character} or \code{vector}) The link between \code{x} and \code{y}. If \code{by} is a \code{character}
string then it is expected to be column of \code{x} and should contain the \code{names} or the indices of the
associated \code{RasterLayer}s in \code{y}. If it is a \code{vector} its length should match the number of rows in \code{x}
and it will be used as if it were a column of \code{x}.}

\item{lng}{(\code{character}) A column of \code{x} that includes the paleolongitudes.}

\item{lat}{(\code{character}) A column of \code{x} that includes the paleolatitudes.}

\item{force}{(\code{character}) If set to \code{"numeric"} the \code{by} argument or the column it points to will be converted to
numeric values, and y will be subsetted with \code{numeric} subscripts of the \code{y} \code{RasterArray}. If set to character, the by column (or vector) will be
forced to \code{character} values and will be used as character subscripts.}
}
\description{
The function takes a set of time-dependent coordinates and extracts the value they point to from associted RasterLayers in a RasterArray.
}
