% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R, R/SpatialStack-base.R
\name{names,RasterArray-method}
\alias{names,RasterArray-method}
\alias{names<-,RasterArray-method}
\alias{names,SpatialStack-method}
\title{Names of one-dimensional RasterArray, SpatialStack or SpatialArray objects.}
\usage{
\S4method{names}{RasterArray}(x)

\S4method{names}{RasterArray}(x) <- value

\S4method{names}{SpatialStack}(x)
}
\arguments{
\item{x}{\code{RasterArray}, \code{SpatialStack} or \code{SpArray} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of names or \code{NULL}.
}
\description{
Get or set the names of one-dimensional RasterArray or SpArray objects
}
\examples{
data(dems)
names(dems)
names(dems)[4] <- "weirdo"
# NULL
data(clim)
names(clim)
}
