% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{ncell,RasterArray-method}
\alias{ncell,RasterArray-method}
\title{Number of cells in a RasterArray object}
\usage{
\S4method{ncell}{RasterArray}(x)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}
}
\description{
The method is inherited from the \code{RasterStack} class.
}
