% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nums}
\alias{nums}
\alias{colnums}
\alias{rownums}
\title{Names as numerics}
\usage{
nums(x)

colnums(x)

rownums(x)
}
\arguments{
\item{x}{Object with names, colnames or rownames attributes.}
}
\value{
Numeric vector.
}
\description{
The set of functions return names of objects directly transformed to numeric values.
}
\examples{
data(dems)
# ages as numerics
nums(dems)
# younger than 20Ma
dems[nums(dems)<20]


}
