% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\name{nvalues}
\alias{nvalues}
\title{The total number of values in a RasterArray object}
\usage{
nvalues(x, ...)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}
}
\description{
The total number of values in a RasterArray object
}
