% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGResponse.R
\name{parse_response}
\alias{parse_response}
\title{Parses a raw httr response into an FGResponse representing an
object of type datatype.  If the httr response contains a parsable
error the function will return an FGErrorResponse instead.}
\usage{
parse_response(response, datatype)
}
\arguments{
\item{response}{The httr response}

\item{datatype}{}
}
\value{
class FGResponse or class FGErrorResponse
}
\description{
Parses a raw httr response into an FGResponse representing an
object of type datatype.  If the httr response contains a parsable
error the function will return an FGErrorResponse instead.
}
