% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{poll_analysis_until_completed}
\alias{poll_analysis_until_completed}
\title{Waits for analysis to complete.}
\usage{
poll_analysis_until_completed(connection, analysis, poll_intervall = 10)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{analysis}{Either the id of an analysis or an FGResponse object containing submitted analysis}

\item{poll_intervall}{The time to wait for a new status update in seconds}
}
\value{
TRUE if the analysis succeeded, otherweise FALSE
}
\description{
Waits for analysis to complete.
}
