% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{poll_app_until_validated}
\alias{poll_app_until_validated}
\title{Waits for the validation of the app to complete.}
\usage{
poll_app_until_validated(connection, app_id, poll_intervall = 10)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{poll_intervall}{The time to wait for a new status update in seconds}

\item{dataset_id}{The app_id of the app OR a FGResponse object}
}
\value{
TRUE if the validation succeeded, otherweise FALSE
}
\description{
Messages and errors are used to show messages. If you need all messages, use fastgenomicsRclient::get_app with the id of this app
}
\examples{
See create_app example
}
