% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{poll_for_upload_to_complete}
\alias{poll_for_upload_to_complete}
\title{Waits for the validation of the dataset to complete.}
\usage{
poll_for_upload_to_complete(connection, dataset_id, poll_intervall = 10)
}
\arguments{
\item{connection}{The connection to be used, call fastgenomicsRclient::connect to obtain one.}

\item{dataset_id}{The id of the dataset, usually starting with dts_***** OR a FGResponse object}

\item{poll_intervall}{The time to wait for a new status update in seconds}
}
\value{
TRUE if the validation succeeded, otherweise FALSE
}
\description{
Messages and errors are used to show messages. If you need all messages, use fastgenomicsRclient::get_dataset with the id of this dataset
}
\examples{
See create_dataset example
}
