% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{projectRaster}
\alias{projectRaster}
\title{Project a RasterArray object}
\usage{
projectRaster(from, to, res, crs, method = "bilinear",
  alignOnly = FALSE, over = FALSE, filename = "", ...)
}
\arguments{
\item{from}{A \code{RasterArray} object to project.}

\item{...}{Arguments passed to the projectRaster() function.}
}
\description{
The uppercase first letter is temporary until the a proper S4 generic of the function is implemented in the \code{raster} package.
}
