% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ramps}
\alias{ramps}
\alias{gradinv}
\alias{ocean}
\alias{terra}
\alias{ipccTemp}
\alias{ipccPrec}
\alias{wet}
\alias{ipccLine}
\alias{ipccRCP}
\alias{showPal}
\title{Colour gradient ramps for default plotting}
\usage{
gradinv(n)

ocean(n)

terra(n)

ipccTemp(n, force = 11)

ipccPrec(n, force = 11)

wet(n)

ipccLine(n = 6)

ipccRCP(n = 4)

showPal(pal = "all")
}
\arguments{
\item{n}{Number of different colors to generate from the palette}

\item{force}{Specify pal when multiple are available. More details to come.}

\item{pal}{A palette name from the lists below}
}
\value{
A function producing a colour gradient ramp.
}
\description{
The object contains functions produced by the \code{colorRampPalette} function.
}
\details{
\code{show.pal} is used to display the available palettes. You can use \code{pal = "all"} or \code{pal=""} if you want to look at all the available palettes. 
You can also view single palettes individually. The following colour palettes are implemented:
\itemize{
\item \code{gradinv()}: inverse heatmap.
\item \code{ocean()}: ocean bathymetrical colours.
\item \code{terra()}: terrestrial topographical colours.
\item \code{ipccTemp()}: gradient from blue to red according to the official IPCC colour palette.
\item \code{ipccPrec()}: gradient from brown to green according to the official IPCC colour palette.
\item \code{wet()}: gradient from white to green to blue. 
\item \code{ipccLine()}: discrete colours for line graphs according to the official IPCC colour palette. 
\item \code{ipccRCP()}: discrete colours for climate scenarios according to the official IPCC colour palette. 
}
}
