% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Generate data report}
\usage{
report(
  inputFile,
  data_refs,
  combine = TRUE,
  enterer_names = NULL,
  output_path = ".",
  output_file = "report.pdf",
  create_dir = TRUE,
  draft = FALSE,
  template = pkg_file("rmarkdown", "templates", "template.tex"),
  skeletonFile = pkg_file("rmarkdown", "templates", "skeleton", "skeleton.Rmd"),
  in_header = NULL,
  quiet = FALSE
)
}
\arguments{
\item{inputFile}{(\code{csv}) path to metadata file. The template can be generated using the \code{\link{create_metadata}} function.}

\item{data_refs}{(\code{data.frame}) data references. See \code{\link{generate_bib}} for more info.}

\item{combine}{(\code{logical}) Should the generated reference file be combined? See \code{\link{generate_bib}} for more info.}

\item{enterer_names}{(\code{character}) names of data enterers to be printed at the end of the report}

\item{output_path}{folder in which to save report}

\item{output_file}{file name of generated report. Must contain ".pdf" extension}

\item{create_dir}{(\code{logical}) should a new directory be created in which the report and related files are stored?}

\item{draft}{(\code{logical}) draft mode allows the editing of the markdown file before the report is generated. Set to \code{FALSE} to generate the file without editing.}

\item{template}{(\code{tex}) LaTex file for the formatting of the report. Uses the in-built template by default.}

\item{skeletonFile}{(\code{Rmd}) Rmd file for the formatting of the report. Uses the in-built template by default.}

\item{in_header}{custom additions to the header, before the \code{\\begin{document}} statement}

\item{quiet}{An option to suppress printing of the pandoc command line.}
}
\description{
Generates a report for data documentation from databases and for meta analyses.
}
\examples{
\donttest{

#Create generic metafile
inputFile <- create_metadata(path=".", edit=FALSE)

# Download references from the Paleobiology Database
refs <- read.csv("https://paleobiodb.org/data1.2/occs/refs.csv?base_name=Scleractinia&occs_created_after=2020-01-01&select=occs&all_records",
# encoding required to allow reading of special characters such as accented characters
encoding="UTF-8")[1:2,]

# Generate report
report(inputFile=inputFile,
data_refs = refs,
output_path = file.path("."),
output_file = "report.pdf",
enterer_names=c("Enterer 1", "Enterer 2"))
}

}
