% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{methods}
\name{resample,RasterArray,ANY-method}
\alias{resample,RasterArray,ANY-method}
\title{Resample a RasterArray object}
\usage{
\S4method{resample}{RasterArray,ANY}(x, y, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{y}{The y argument of the \code{\link[raster]{resample}} function.}

\item{...}{arguments passed to the \code{\link[raster]{resample}} function.}
}
\value{
A resampled \code{RasterArray} class object.
}
\description{
The method is inherited from the \code{RasterStack} class.
}
\examples{
data(dems)
template <- raster(res=5)
resampled <- resample(dems, template)
}
