% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{revoke_share}
\alias{revoke_share}
\title{Revoke a share for a recipient}
\usage{
revoke_share(connection, share_id, recipient_email)
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}

\item{share_id}{The id of the share}

\item{recipient_email}{The email of the recipient}
}
\value{
A FGResponse object
}
\description{
If a recipient is unknown, a NotFound Message will be generated
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
share <- create_share(conn, "dts_0f91c4f034e1471fa25f1078d23e1a97", message = "no msg", license = "license", send_to = list("a2@example.com"))
revoke_share(conn, share@Id, "a2@example.com")
}
