% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{rotate}
\alias{rotate}
\alias{rotate,RasterArray-method}
\title{Rotate a \code{\link[chronosphere:RasterArray-class]{RasterArray}}) object}
\usage{
\S4method{rotate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{(\code{\link[chronosphere:RasterArray-class]{RasterArray}}) Object.}

\item{...}{(Additional arguments passed to the \code{\link[raster]{rotate}}) function.}
}
\value{
A (\code{\link[chronosphere:RasterArray-class]{RasterArray}})-class object.
}
\description{
The method is inherited from the \code{\link[raster:raster]{RasterStack}} class.
}
