% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{rotationModels}
\alias{rotationModels}
\title{Available rotation models}
\usage{
rotationModels()
}
\description{
This function outputs a vector with the available rotation models for the online/remote
reconstruction.
}
\details{
The available reconstruction models are:
\itemize{
 \item "SETON2012" (Seton et al., 2012) for coastlines and plate polygons (0 - 200 Ma).
 \item "MULLER2016" (Muller et al., 2016) for coastlines and plate polygons(0 - 230 Ma).
 \item "GOLONKA" (Wright et al. 2013) for coastlines only (0 - 550 Ma). 
 \item "PALEOMAP" (Scotese and Wright, 2018) for coastlines and plate polygons (0 - 750 Ma). 
 \item "MATTHEWS2016" (Matthews et al., 2016) for coastlines and plate polygons (0 - 410 Ma). 
}
}
