% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{rownames,RasterArray-method}
\alias{rownames,RasterArray-method}
\alias{rownames<-,RasterArray-method}
\title{Row names of two-dimensional RasterArray or SpArray objects}
\usage{
\S4method{rownames}{RasterArray}(x)

\S4method{rownames}{RasterArray}(x) <- value
}
\arguments{
\item{x}{\code{RasterArray} or \code{SpArray} object.}

\item{value}{\code{character} vector.}
}
\description{
Get or set the row names of two-dimensional RasterArray or SpArray objects
}
\examples{
# an example
}
