% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{This function orders an analysis to run on FASTGenomics.  Aside
from the dataset id and the workflow id you will need to specify a
title, abstract and optionally a description.}
\usage{
run_analysis(connection, dataset_id, workflow_id, title, abstract,
  description = "")
}
\arguments{
\item{connection}{The connection to be used, call \code{\link{connect}} to obtain one.}
}
\value{
class FGResponse
}
\description{
This function orders an analysis to run on FASTGenomics.  Aside
from the dataset id and the workflow id you will need to specify a
title, abstract and optionally a description.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Bearer ey...")
analysis <- fastgenomicsRclient::run_analysis(connection, "dts_xxxx", "wf_xxx", title="Some title", abstract="Some abstract")
status <- fastgenomicsRclient::poll_analysis_until_validated(connection, analysis)
}
