% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{save_personal_access_token}
\alias{save_personal_access_token}
\title{Store a Personal Access Token (PAT) on your system}
\usage{
save_personal_access_token(base_url, email)
}
\arguments{
\item{base_url}{The url of the instance, e.g. https://fastgenomics.org/}

\item{email}{The email address of your account}
}
\value{
Nothing
}
\description{
You will be asked for a PAT, please generate it at: <base_url>/ids/Manage/ManagePats
The PAT is stored securely in a keyring, which can only be accessed by your user.
WARNINIG: NEVER store your PAT in your R history or in Source Control.
Anybody who has the PAT can take any action on FASTGenomics you can make.
More Questions? Read our \href{https://github.com/FASTGenomics/fastgenomics-docs/blob/master/doc/api/authorization_guide.md}{in depth authorization Guide}
}
\examples{
fastgenomicsRclient::save_personal_access_token("https://fastgenomics.org/", "user@example.com")
}
