% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{shaper}
\alias{shaper}
\title{Code snippets defining ranges based on points located on a plot}
\usage{
shaper(f = "p", n = 2, round = 2, ...)
}
\arguments{
\item{f}{(\code{character}) A single letter value specifying for which function's arugment format you want to get parameters. \code{"p"} is for \code{\link[graphics]{plot}}, \code{"r"} is for \code{\link[graphics]{rect}}, \code{"s"} is for \code{\link[graphics]{segments}}. \code{"e"} returns a call to create an \code{\link[raster]{extent}} class object from the package \code{raster}. \code{"m"} will return code to define a 2 column matrix.}

\item{n}{(\code{integer}) The number of points to request.}

\item{round}{(\code{integer}) Number of digits to round to, can be two values, first is for \code{x} second for \code{y}.}

\item{...}{arguments passed to the \code{\link[graphics]{locator}} fucntion}
}
\description{
The function returns snippets of code that you can paste in your script after you select points on a plot. Useful for defining areas on a map. The default methods assume that you will first click in the bottom left and then in the bottom right corner.
}
