% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_climate_change.R
\name{sp_linear_rate}
\alias{sp_linear_rate}
\title{Calculate the rate of climate change over a given time period for a specified unit of time}
\usage{
sp_linear_rate(x, period = NULL, t = 1)
}
\arguments{
\item{x}{is a raster object of the variable of interest, with each layer representing time}

\item{period}{is a vector contains the time intervals corresponding to each raster layers, with negative values representing time before present.}

\item{t}{is the unit of time considered, e.g. if the rate per 5 myr is wanted, then t=5.}
}
\value{
RasterLayer
}
\description{
\code{sp_linear_rate} calculates the spatial gradient of climatic change, given a variable of interest over a period of time.
}
\details{
Based on Burrows et al. (2011)
}
\references{
Loarie, S. R., P. B. Duffy, H. Hamilton, G. P. Asner, C. B. Field, and D. D. Ackerly. 2009. The velocity of climate change. Nature 462:1052–5.

Burrows, M. T., D. S. Schoeman, L. B. Buckley, P. Moore, E. S. Poloczanska, K. M. Brander, C. Brown, J. F. Bruno, C. M. Duarte, B. S. Halpern, J. Holding, C. V Kappel, W. Kiessling, M. I. O’Connor, J. M. Pandolfi, C. Parmesan, F. B. Schwing, W. J. Sydeman, and A. J. Richardson. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science (New York, N.Y.) 334:652–5.
}
