% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_climate_change.R
\name{spatial_gradient}
\alias{spatial_gradient}
\title{Calculate the spatial gradient of climatic change}
\usage{
spatial_gradient(x)
}
\arguments{
\item{x}{is a single-layer raster containing the variable of interest
averaged over a particular time period.}
}
\value{
RasterLayer
}
\description{
\code{spatial_gradient} calculates the spatial gradient of climatic change, given a variable of interest over a period of time.
}
\details{
Based on Burrows et al. (2011)
}
\references{
Loarie, S. R., P. B. Duffy, H. Hamilton, G. P. Asner, C. B. Field, and D. D. Ackerly. 2009. The velocity of climate change. Nature 462:1052–5.

Burrows, M. T., D. S. Schoeman, L. B. Buckley, P. Moore, E. S. Poloczanska, K. M. Brander, C. Brown, J. F. Bruno, C. M. Duarte, B. S. Halpern, J. Holding, C. V Kappel, W. Kiessling, M. I. O’Connor, J. M. Pandolfi, C. Parmesan, F. B. Schwing, W. J. Sydeman, and A. J. Richardson. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science (New York, N.Y.) 334:652–5.
}
