% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-base.R
\name{stack,VectorSpatialClasses-method}
\alias{stack,VectorSpatialClasses-method}
\title{Stacking method for the \code{\link{SpatialStack}} objects}
\usage{
\S4method{stack}{VectorSpatialClasses}(x, ...)
}
\arguments{
\item{x}{\code{SpatialPoints},\code{SpatialPointsDataFrame},\code{SpatialLines},\code{SpatialLinesDataFrame},\code{SpatialPolygons},\code{SpatialPolygonsDataFrame}, object.}
}
\value{
A \code{\link{RasterArray}} class object.
}
\description{
The function alls a \code{\link{RasterArray}}-like stacking of Spatial* objects.
}
