% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-base.R
\name{[,SpatialStack,ANY,ANY-method}
\alias{[,SpatialStack,ANY,ANY-method}
\title{Indexing to extract subsets of a SpatialStack object}
\usage{
\S4method{[}{SpatialStack,ANY,ANY}(x, i, drop = TRUE)
}
\arguments{
\item{x}{\code{SpatialStack} object.}

\item{i}{subscript of vector-like subsetting.}

\item{drop}{\code{logical} in case the result of subsetting is a single element, should the \code{SpatialStack} wrapper be dropped and the element be reduced to a single \code{Spatial}?}
}
\value{
A \code{SpatialStack} or \code{Spatial} class object.
}
\description{
The single and double bracket subsetting is identical in the case of SpatiaStacks.
}
