% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-subset.R
\docType{methods}
\name{[[,RasterArray,ANY,ANY-method}
\alias{[[,RasterArray,ANY,ANY-method}
\title{Indexing to extract RasterLayers of a RasterArray object}
\usage{
\S4method{[[}{RasterArray,ANY,ANY}(x, i, drop = TRUE)
}
\arguments{
\item{x}{\code{RasterArray} object.}

\item{i}{subscript of the first dimension(rows) or vector-like subsetting.}

\item{drop}{\code{logical} should the \code{RasterStack} be dropped and the element be reduced to a single \code{RasterLayer}?}
}
\value{
The function returns either a single \code{RasterLayer}, a \code{RasterStack} or a list of \code{RasterLayers}.
}
\description{
Double bracket \code{'[['} refers to layers' name in the \code{RasterStack} of the \code{RasterArray}. Use single brackets to extract elements based on their position in the \code{RasterArray}.
}
