% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-subset.R
\name{[[,SpatialStack,ANY,ANY-method}
\alias{[[,SpatialStack,ANY,ANY-method}
\title{Indexing to extract Spatial items from a  SpatialStack object}
\usage{
\S4method{[[}{SpatialStack,ANY,ANY}(x, i, drop = TRUE)
}
\arguments{
\item{x}{\code{SpatialStack} object.}

\item{i}{subscript of vector-like subsetting.}

\item{drop}{\code{logical} should the \code{SpatialStack} be dropped and the element be reduced to a single \code{Spatial} object?}
}
\value{
A \code{Spatial} or \code{SpatialStack} class object.
}
\description{
The single and double bracket subsetting is identical in the case of SpatialStacks.#'
}
