% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-subset.R
\docType{methods}
\name{[[<-,RasterArray,ANY,ANY-method}
\alias{[[<-,RasterArray,ANY,ANY-method}
\title{Replace RasterLayers in a RasterArray object}
\usage{
\S4method{[[}{RasterArray,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{\code{RasterArray} object.}

\item{i}{subscript of layers to replace.}

\item{value}{\code{character} vector.}
}
\description{
Double bracket \code{'[['} refers to layers' name in the \code{RasterStack} of the \code{RasterArray}. Use single brackets to replace elements based on their position in the \code{RasterArray}.
}
\examples{
# an example
}
