% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-subset.R
\docType{methods}
\name{[<-,RasterArray,ANY,ANY,RasterLayer-method}
\alias{[<-,RasterArray,ANY,ANY,RasterLayer-method}
\title{Replace RasterLayers in a RasterArray object}
\usage{
\S4method{[}{RasterArray,ANY,ANY,RasterLayer}(x, i, j, ...) <- value
}
\arguments{
\item{x}{\code{RasterArray} object.}

\item{value}{A \code{RasterLayer} or \code{RasterArray} object.}
}
\description{
Replace RasterLayers in a RasterArray object
}
\examples{
# an example
}
