% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-subset.R
\name{subset,SpatialStack-method}
\alias{subset,SpatialStack-method}
\title{Subset a SpatialStack object}
\usage{
\S4method{subset}{SpatialStack}(x, i, drop = TRUE)
}
\arguments{
\item{x}{\code{SpatialStack} object.}

\item{i}{subscript ofvector-like subsetting.}

\item{drop}{\code{logical} in case the result of subsetting is a single element, should the \code{SpatialStack} wrapper be dropped and the element be reduced to a single \code{RasterLayer}?}
}
\value{
A \code{Spatial} or \code{SpatialStack} class object.
}
\description{
Extract subsets of SpatialStack class object similarly to a RasterStack
}
\examples{
# stack of the paleocoast coastlines
data(coastlines)
spstack <- coastlines@stack
subset(spstack, "X5Ma_CS_v7")
}
