% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\docType{methods}
\name{summary,RasterArray-method}
\alias{summary,RasterArray-method}
\title{Summary}
\usage{
\S4method{summary}{RasterArray}(object)
}
\arguments{
\item{object}{a \code{RasterArray} class object.}
}
\description{
Summarizes a \code{RasterArray} class object.
}
\details{
The method is inherited from the \code{RasterStack} class.
}
