% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-base.R
\docType{data}
\name{t}
\alias{t}
\alias{t,XArray-method}
\title{Transpose a \code{\link{RasterArray}} or \code{\link{SpatialArray}} object}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
t

\S4method{t}{XArray}(x)
}
\arguments{
\item{x}{A \code{\link{RasterArray}} or \code{\link{SpatialArray}} class object.}
}
\value{
A \code{\link{RasterArray}} or \code{\link{SpatialArray}} class object.
}
\description{
Transpose a \code{\link{RasterArray}} or \code{\link{SpatialArray}} object
}
\examples{
data(dems)
t(dems)
data(coastlines)
t(coastlines)
}
\keyword{datasets}
