% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\docType{methods}
\name{t,RasterArray-method}
\alias{t,RasterArray-method}
\title{Transpose a RasterArray object}
\usage{
\S4method{t}{RasterArray}(x)
}
\arguments{
\item{x}{A RasterArray class object.}
}
\description{
Transpose a RasterArray object
}
\examples{
data(dems)
t(dems)
data(clim)
t(clim)
}
