% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruction.R
\name{validCoords}
\alias{validCoords}
\title{Validate coordinates}
\usage{
validCoords(x, invalid = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{coordinates to be checked. Can be a \code{vector}, \code{matrix} or \code{data.frame}.}

\item{invalid}{(\code{logical}) Return vector of indexes of incomplete/invalid coordinates?}

\item{verbose}{(\code{logical}) Should console feedback be printed?}
}
\description{
Utility function to check if provided coordinates are valid
}
