% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_climate_change.R
\name{vocc}
\alias{vocc}
\title{Calculate the velocity of climate change over a given set of years}
\usage{
vocc(x, period, t = 1, truncate = TRUE, thr = 500)
}
\arguments{
\item{x}{is the climate data as a raster object.}

\item{period}{is a numeric vector specifying the time interval over which the velocity of
climate change will be calculated. Should correspond to the name of the raster layers 
although time intervals before present are given as negative values.}

\item{truncate}{logical expression indicating whether minimum and maximum
values should be constrained to a certain threshold. Default is TRUE.}

\item{thr}{is the threshold to contrain the minimum and maximum values of the climate velocity.}
}
\value{
a two-layer raster containing 'velocity_mag' and 'velocity_angle',
the magnitude and direction of the velocity of climate change, respectively.
}
\description{
Calculate the velocity of climate change over a given set of years
}
\references{
Loarie, S. R., P. B. Duffy, H. Hamilton, G. P. Asner, C. B. Field, and D. D. Ackerly. 2009. The velocity of climate change. Nature 462:1052–5.

Burrows, M. T., D. S. Schoeman, L. B. Buckley, P. Moore, E. S. Poloczanska, K. M. Brander, C. Brown, J. F. Bruno, C. M. Duarte, B. S. Halpern, J. Holding, C. V Kappel, W. Kiessling, M. I. O’Connor, J. M. Pandolfi, C. Parmesan, F. B. Schwing, W. J. Sydeman, and A. J. Richardson. 2011. The pace of shifting climate in marine and terrestrial ecosystems. Science (New York, N.Y.) 334:652–5.
}
