% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-xattrib.R
\name{ncol,XArray-method}
\alias{ncol,XArray-method}
\alias{nrow,XArray-method}
\title{Number of columns and rows of a \code{\link{RasterArray}} or \code{\link{SpatialArray}}}
\usage{
\S4method{ncol}{XArray}(x)

\S4method{nrow}{XArray}(x)
}
\arguments{
\item{x}{A \code{\link{RasterArray}} or \code{\link{SpatialArray}}  class object.}
}
\value{
A \code{numeric} value of the number of columns and rows.
}
\description{
Unlike the \code{ncol} and \code{nrow} functions of the raster package (\code{\link[raster]{ncell}}), this function returns the number of columns and rows of the \code{RasterArray} container, rather than the dimensions of the contained \code{RasterLayer}s.
}
\examples{
data(coasts)
ncol(coasts)
nrow(coasts)
}
