% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\name{as.data.frame.RasterArray}
\alias{as.data.frame.RasterArray}
\title{S3-type method for RasterArray allowing View(), head() and tail() to work.}
\usage{
as.data.frame.RasterArray(from, to = "data.frame")
}
\description{
S3-type method for RasterArray allowing View(), head() and tail() to work.
}
