% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-cast.R
\name{as.RasterArray}
\alias{as.RasterArray}
\alias{as.RasterArray,RasterLayer-method}
\alias{as.RasterArray,RasterStack-method}
\alias{as.RasterArray,RasterBrick-method}
\title{Convert Raster* objects to RasterArrays}
\usage{
as.RasterArray(from)

\S4method{as.RasterArray}{RasterLayer}(from)

\S4method{as.RasterArray}{RasterStack}(from)

\S4method{as.RasterArray}{RasterBrick}(from)
}
\arguments{
\item{from}{Object to be converted.}
}
\value{
A \code{RasterArray} class object.
}
\description{
The function converts \code{RasterLayer}, \code{RasterStack} and \code{RasterBrick} objects to \code{RasterArray} class objects.
}
\examples{
data(dems)
recent <- dems[1]
# convert RasterLayer to RasterArray
ra <- as.RasterArray(recent)
}
