% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialArray-cast.R
\name{as.SpatialArray}
\alias{as.SpatialArray}
\title{Convert Spatial* objects to SpatialArrays}
\usage{
as.SpatialArray(from)
}
\arguments{
\item{from}{Object to be converted.}
}
\value{
A \code{SpatialArray} class object.
}
\description{
The function converts \code{Spatial*}objects to \code{\link{SpatialArray}}-class objects.
}
\examples{
# data(coastlines)
# recent <- coastlines[1]
# # convert Spatial* to as.SpatialArray
# sa <- as.SpatialArray(recent)
}
