% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{chronosphere}
\alias{chronosphere}
\title{Earth System History Variables}
\description{
The purpose of the 'chronosphere' project is to facilitate spatially explicit analyses of (paleo)environmental/ecological research. The package serves as a gateway to plate tectonic reconstructions, deep time global climate model results as well as fossil occurrence datasets such as the Paleobiology Database \url{https://www.paleobiodb.org/} and the PaleoReefs Database \url{https://www.paleo-reefs.pal.uni-erlangen.de/}. Environmental data stored on a remote server can be downloaded and imported directly to the R environment. Query functions to the GPlates (\url{http://www.gplates.org/}) desktop application or the GPlates Web Service (\url{https://gws.gplates.org/}) allow users to reconstruct coordinates, static plates, and Spatial objects. A wrapper class \code{\link{RasterArray}} is implemented around the \code{\link[raster:raster]{RasterLayer}} class, allowing the organization of spatially explicit raster data in \code{n}-dimensional arrays. The project is developed under the umbrella of the DFG (Deutsche Forschungsgemeinschaft) Research Unit TERSANE2 (For 2332, TEmperature Related Stressors in ANcient Extinctions).
}
\details{
This is still the Beta version. As is R, This is free software and comes with ABSOLUTELY NO WARRANTY. Nevertheless, notes about found bugs and suggestions are more than welcome.
}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com) and Nussaibah B. Raja
}
