% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clim}
\alias{clim}
\title{Annual mean temperature (bio1) and precipitation (bio12) variables between 2001 and 2010 from the CHELSA model}
\format{A RasterArray with 20 layers, 10 rows and two columns.}
\source{
DEMS
}
\usage{
clim
}
\description{
Currently this is just a dummy made from the DEMS.
}
\keyword{datasets}
