% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-cast.R
\name{as}
\alias{as}
\title{Coerce RasterLayer, RasterStack and RasterBrick object to a RasterArray}
\arguments{
\item{from}{Object to be coerced.}
}
\description{
The function coerces RasterLayer, RasterStack and RasterBrick object to a RasterArray.
}
