% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-combine.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,RasterLayer-method}
\alias{combine,RasterArray-method}
\title{Combine RasterLayers and one-dimensional RasterArrays}
\usage{
combine(x, ...)

\S4method{combine}{RasterLayer}(x, ...)

\S4method{combine}{RasterArray}(x, ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterArray} objects to combine.}
}
\description{
Methods sequences that start with an NA do not yet work.
}
\examples{
data(demo)
a <- combine(demo[1], demo[2])
}
