% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-combine.R
\name{combine}
\alias{combine}
\alias{combine,RasterLayer-method}
\alias{combine,RasterArray-method}
\title{Combine RasterLayers and one-dimensional RasterArrays}
\usage{
combine(x, ...)

\S4method{combine}{RasterLayer}(x, ...)

\S4method{combine}{RasterArray}(x, ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterArray} object to combine.}

\item{...}{additional objects to combine.}
}
\value{
A \code{RasterArray} class object.
}
\description{
Methods sequences that start with an NA do not yet work.
}
\examples{
data(dems)
a <- combine(dems[1], dems[2])
}
