% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Get a FastGenomics connection object to a specific FASTGenomics instance}
\usage{
connect(base_url, email)
}
\arguments{
\item{base_url}{The url of the instance, e.g. https://fastgenomics.org/}

\item{email}{The email address of your account}
}
\value{
a connection object
}
\description{
WARNINIG: NEVER commit passwords or Tokens to a GIT repository or share them in any way! The token can be used to do any action on your behalf.
In the background, your PAT is used.
}
\examples{
fastgenomicsRclient::save_personal_access_token("https://fastgenomics.org/", "user@example.com")
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "user@example.com")
}
