% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect}
\alias{connect}
\title{Get a FastGenomics connection object to a specific FASTGenomics instance}
\usage{
connect(base_url, email)
}
\arguments{
\item{base_url}{The url of the instance, e.g. https://fastgenomics.org/}

\item{bearer_token}{The API token of your user, you can get it here: {base_url}/ids/Account/ApiTokenLogin}
}
\value{
a connection object
}
\description{
Note that our tokens have an expiration date, e.g. after 6 hours you will need to obtain a new Bearer token an create a new connection object.
WARNINIG: NEVER commit passwords or Tokens to a GIT repository or share them in any way! The token can be used to do any action on your behalf.
}
\examples{
connection <- fastgenomicsRclient::connect("https://fastgenomics.org/", "Beaer ey...")
}
