% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-seurat.R
\name{create_dataset_from_seurat}
\alias{create_dataset_from_seurat}
\title{Uploads a seurat dataset.  For argument description see \code{\link{create_dataset_df}}.}
\usage{
create_dataset_from_seurat(connection, seurat_obj, ...)
}
\description{
Uploads a seurat dataset.  For argument description see \code{\link{create_dataset_df}}.
}
