% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{create_metadata}
\alias{create_metadata}
\title{Create metadata file to generate report}
\usage{
create_metadata(path, edit = TRUE, overwrite = FALSE, return_path = TRUE)
}
\arguments{
\item{path}{path where to create file}

\item{edit}{open file after creation for edits}

\item{overwrite}{(\code{logical}) Should the file be overwritten if already exists?}

\item{return_path}{(\code{logical}) return the file path of the create file}
}
\description{
Create metadata file to generate report
}
\examples{
\dontrun{
create_metadata(path=".", edit=FALSE)
}
}
