% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\name{dim,RasterArray-method}
\alias{dim,RasterArray-method}
\title{Dimensions of RasterArray objects}
\usage{
\S4method{dim}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}
}
\value{
A \code{numeric} vector.
}
\description{
The function returns the dimensions of the array in which \code{RasterLayer}s are organized.
}
\examples{
data(dems)
dim(dems)
data(clim)
dim(clim)
}
