% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\name{dimlayer}
\alias{dimlayer}
\title{Dimensions of a RasterLayers in a RasterArray}
\usage{
dimlayer(x, ...)
}
\arguments{
\item{x}{A \code{RasterLayer} class object.}
}
\description{
The funcion will return the dimensions RasterLayers
}
