% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-raster.R
\name{disaggregate,RasterArray-method}
\alias{disaggregate,RasterArray-method}
\title{Disaggregate raster cells in a RasterArray object}
\usage{
\S4method{disaggregate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{RasterArray} class object.}

\item{...}{arguments passed to the \code{\link[raster]{disaggregate}} function.}
}
\value{
A disaggregated \code{RasterArray} class object.
}
\description{
The method is inherited from the \code{RasterStack} class.
}
\examples{
data(dems)
disagg <- disaggregate(dems, 3)

}
