% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{fetch}
\alias{fetch}
\title{Data fetching}
\usage{
fetch(
  dat = NULL,
  var = NULL,
  ver = NULL,
  res = NULL,
  ext = NULL,
  class = NULL,
  item = NULL,
  datadir = NULL,
  verbose = TRUE,
  call = FALSE,
  call.expr = FALSE,
  attach = TRUE,
  ...
)
}
\arguments{
\item{dat}{(\code{character}) The dataset to get variables from.}

\item{var}{(\code{character}) Vector of variable names to get.}

\item{ver}{(\code{character}) The version of the variable. Defaults to \code{NULL}, which will download the latest available version. We have to create a data table, which should be part of the package. This has to be searched for valid argument combinations. Right this is just a folder with a date.}

\item{res}{(\code{character} or \code{numeric}) The resolution of raster layers. This has to be the same for all RasterLayers that make up the variable.}

\item{ext}{(\code{character}) File extension of the used data file.}

\item{class}{(\code{character}) Class of the returned object, if not the default.}

\item{item}{(\code{numeric}) The item ID that is to be downloaded. This setting overrides all other identifiers.}

\item{datadir}{(\code{character}) Directory where downloaded files are kept. Individual layers will be looked up from the directory if this is given, and will be downloaded if they are not found. The default \code{NULL} option will download data to a temporary directory that exists only until the R session ends.}

\item{verbose}{(\code{logical}) Should console feedback during download be displayed?}

\item{call}{(\code{logical}) If set to \code{TRUE} the function call is returned instead of the object.}

\item{call.expr}{(\code{logical}) If \code{call} is set to \code{TRUE}, then should the call be returned as an \code{expression} (\code{TRUE}) or a message (\code{FALSE})?}

\item{attach}{(\code{logical}) If the item has required packages, should these be attached?}

\item{...}{Arguments passed to variable-specific loading functions.}
}
\value{
An object that matches the 'type' field of the varibles in the output of the \code{\link{datasets}} function.
}
\description{
Function to download and attach variables in the \code{chronosphere} package
}
\details{
Use the function \code{\link{datasets}} to find available variables.
}
\examples{
# An actual download call
# a <- fetch(dat="paleomap", var="dem")
# call repetition
fetch(dat="paleomap", var="dem", call=TRUE)
# A locally-present object, in package's directory
a <- fetch(dat="paleomap", var="model", 
  datadir=system.file("extdata", package="chronosphere"))
}
