% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{fetch}
\alias{fetch}
\title{Data layer fetching}
\usage{
fetch(dat, var, res = 1, ver = NULL, datadir = NULL)
}
\arguments{
\item{dat}{The dataset to get variables from.}

\item{var}{Vector of variable names to get.}

\item{res}{The resolution of the layers. This has to be the same for all layers.}

\item{ver}{The version of the variable. Defaults to \code{NULL}, which will download the latest available version. We have to create a data table, which should be part of the package. This has to be searched for valid argument combinations. Right this is just a folder with a date.}

\item{datadir}{Directory where downloaded files are kept. Individual layers will be looked up from the directory if this is given, and will be downloaded if they are not found. The default \code{NULL} option will download data to a temporary directory that exists only until the R session ends.}
}
\description{
Function to download and attach variables in the \code{chronosphere} package
}
\details{
The function is implemented for a single variable, which can be downloaded and attached efficiently (dem). There are a lot of things to figure out.
Version 2.
}
\examples{
	a <- fetch(dat="paleomap", var="dem")
	b <- fetch(dat="paleomap", var="paleoatlas", res=0.1)
}
