% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGDatasetUploadParameters.R
\name{FGDatasetUploadParameters}
\alias{FGDatasetUploadParameters}
\title{Get the optional parameters to create a dataset}
\usage{
FGDatasetUploadParameters(license = "", web_link = "", citation = "",
  technology = "", current_normalization_status = "",
  cell_metadata = "", gene_metadata = "")
}
\arguments{
\item{license}{The license to be used. If the data are owned by yourself, you can use https://creativecommons.org/choose/ to choose a license. If not, you have to find the license used by the dataset, e.g. talk to your supervisor.}

\item{web_link}{The website of your dataset.}

\item{citation}{How to cite this dataset}

\item{technology}{The technology used to obtain this dataset. call \code{\link{get_valid_technologies}} to get a list of already known technologies or freely choose a name.}

\item{current_normalization_status}{The current_normalization_status used to obtain this dataset. call \code{\link{get_valid_current_normalization_status}} to get a list of already known normalization techniques or freely choose a name.}

\item{cell_metadata}{The path to your cell metadata file OR a dataframe, for valid formats see \href{https://github.com/FASTGenomics/fastgenomics-docs/blob/master/doc/api/dataset_api.md}{here}.}

\item{gene_metadata}{The path to your gene metadata OR a dataframe, for valid formats see \href{https://github.com/FASTGenomics/fastgenomics-docs/blob/master/doc/api/dataset_api.md}{here}.}
}
\value{
class FGDatasetUploadParameters
}
\description{
If you provide cell or gene metadata via a dataframe, please note that the dataframe is stored on disk in the tempdir of the current R session. As this might get cleaned on exit of R, be cautious when using this from a stored R session.
}
\examples{
optional <- fastgenomicsRclient::FGDatasetUploadParameters(
                                        license ="MIT",
                                        technology = "Smart-Seq",
                                        web_link="https://example.com",
                                        notes="This is a TEST",
                                        citation="FG et al",
                                        batch_column="sample",
                                        current_normalization_status="Counts",
                                        cell_metadata="./cell_metadata.tsv",
                                        gene_metadata="./gene_metadata.tsv"  )
}
