% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Documentation page of a series (Remote server under construction)}
\usage{
info(src, ser = NULL)
}
\arguments{
\item{src}{(\code{character}) Object downloaded with \code{\link{fetch}} or the source identifier string.}

\item{ser}{(\code{character}) In case \code{src} is \code{character}, the series identifier.}
}
\value{
The function has no return value.
}
\description{
Documentation page of a series (Remote server under construction)
}
