% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-xattrib.R
\name{layers}
\alias{layers}
\alias{layers,XArray-method}
\title{Names of \code{\link{RasterArray}} or \code{\link{SpatialArray}} Layers in the stack}
\usage{
layers(x, ...)

\S4method{layers}{XArray}(x)
}
\arguments{
\item{x}{A \code{\link{RasterArray}} or \code{\link{SpatialArray}} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{character} vector of names.
}
\description{
Names of \code{\link{RasterArray}} or \code{\link{SpatialArray}} Layers in the stack
}
\examples{
# names of layers in the stack
data(dems)
layers(dems)
}
