% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{layers}
\alias{layers}
\alias{layers,RasterArray-method}
\title{Names of RasterArray's Layers in the stack}
\usage{
layers(x, ...)

\S4method{layers}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\description{
Names of RasterArray's Layers in the stack
}
\examples{
# names of layers in the stack
data(dems)
layers(dems)
}
