% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mapplot}
\alias{mapplot}
\title{Pseudo generic function to plot maps of different object classes}
\usage{
mapplot(x, legend = FALSE, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{legend}{(\code{logical}) Triggers whether the legend of a RasterLayer would be plotted.}
}
\description{
This function plots the different paleo
}
