% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{mapplot}
\alias{mapplot}
\alias{mapplot,RasterLayer-method}
\alias{mapplot,RasterStack-method}
\alias{mapplot,RasterArray-method}
\title{Wrapper function to plot maps of different classes}
\usage{
mapplot(x, ...)

\S4method{mapplot}{RasterLayer}(x, legend = FALSE, col = gradinv(255),
  ...)

\S4method{mapplot}{RasterStack}(x, col = gradinv(255), ...)

\S4method{mapplot}{RasterArray}(x, col = gradinv(255), ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{legend}{(\code{logical}) Triggers whether the legend of a RasterLayer would be plotted.}

\item{col}{\code{character} Color or color scheme of the plot.}
}
\description{
This function plots Raster and sp-type objects.
}
