% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{mapplot}
\alias{mapplot}
\alias{mapplot,RasterLayer-method}
\alias{mapplot,RasterStack-method}
\alias{mapplot,RasterArray-method}
\alias{mapplot,SpatialPolygonsDataFrame-method}
\alias{mapplot,SpatialPolygons-method}
\title{Wrapper function to plot maps of different classes}
\usage{
mapplot(x, ...)

\S4method{mapplot}{RasterLayer}(x, col = gradinv(255), axes = FALSE,
  box = FALSE, legend = FALSE, legend.title = NULL, ...)

\S4method{mapplot}{RasterStack}(x, col = gradinv(255), ...)

\S4method{mapplot}{RasterArray}(x, col = gradinv(255), rgb = FALSE,
  legend = FALSE, axes = FALSE, box = FALSE, ask = TRUE,
  nplots = 6, legend.title = NULL, plot.title = NULL, ...)

\S4method{mapplot}{SpatialPolygonsDataFrame}(x, col = "lightgrey", ...)

\S4method{mapplot}{SpatialPolygons}(x, col = "lightgrey", ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{col}{\code{character} Color or color scheme of the plot. See \code{?ramps} for available palettes.}

\item{legend}{(\code{logical}) Triggers whether the legend of a RasterLayer would be plotted.}
}
\description{
This function plots Raster and sp-type objects.
}
