% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{names,RasterArray-method}
\alias{names,RasterArray-method}
\alias{names<-,RasterArray-method}
\title{Names of one-dimensional RasterArray or SpArray objects.}
\usage{
\S4method{names}{RasterArray}(x)

\S4method{names}{RasterArray}(x) <- value
}
\arguments{
\item{x}{\code{RasterArray} or \code{SpArray} object.}

\item{value}{\code{character} vector.}
}
\description{
Get or set the names of one-dimensional RasterArray or SpArray objects
}
\examples{
data(dems)
names(dems)
names(dems)[4] <- "weirdo"
# NULL
data(clim)
names(clim)
}
