% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-xattrib.R
\docType{methods}
\name{nvalues}
\alias{nvalues}
\alias{nvalues,RasterArray-method}
\title{The total number of values in a RasterArray object}
\usage{
nvalues(x, ...)

\S4method{nvalues}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\description{
The total number of values in a RasterArray object
}
\examples{
data(dems)
nvalues(dems)
}
