% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plot,RasterArray,missing-method}
\alias{plot,RasterArray,missing-method}
\title{Shorthand for the plotting RasterArray objects}
\usage{
\S4method{plot}{RasterArray,missing}(x, y, ...)
}
\arguments{
\item{x}{A RasterArray class object.}

\item{y}{Not implemented yet.}

\item{...}{Arguments passed to the \code{\link{mapplot}} function.}
}
\description{
The plot(), method executes the \code{\link{mapplot}} function on the RasterArray object.
}
\examples{
data(dems)
plot(dems)
}
