% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\name{proxy}
\alias{proxy}
\alias{proxy,RasterArray-method}
\title{The proxy of a RasterArray object}
\usage{
proxy(x, ...)

\S4method{proxy}{RasterArray}(x)
}
\arguments{
\item{x}{(\code{RasterArray}  focal object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{vector}, \code{matrix} or \code{array} of characters representing the \code{RasterArray} structure.
}
\description{
This function returns an object that symbolizes the structure of layers in the \code{RasterArray}.
}
\details{
The \code{proxy} method wraps the names of layers in the stack using the \code{index} slot of the \code{RasterArray}.
}
\examples{
data(dems)
proxy(dems)

data(clim)
proxy(clim)
}
