% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-base.R
\name{proxy}
\alias{proxy}
\title{The proxy of a RasterArray or SpArray object}
\usage{
proxy(x, ...)
}
\arguments{
\item{x}{(\code{RasterArray} or \code{SpArray}) focal object.}
}
\description{
This function extracts the @index slot of a RasterArray or SpArray-class object
}
