% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gradinv}
\alias{gradinv}
\title{Colour gradient ramps for default plotting}
\usage{
gradinv(n)
}
\description{
The object contains functions produced by the \code{colorRampPalette} function.
}
\details{
The following ramps are implemented:
gradinv(): inverse heatmap.
}
