% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-subset.R
\name{[<-,SpatialStack,character,ANY,VectorSpatialClasses-method}
\alias{[<-,SpatialStack,character,ANY,VectorSpatialClasses-method}
\alias{[<-,SpatialStack,logical,ANY,VectorSpatialClasses-method}
\alias{[<-,SpatialStack,numeric,ANY,VectorSpatialClasses-method}
\alias{[<-,SpatialStack,character,ANY,SpatialStack-method}
\alias{[<-,SpatialStack,logical,ANY,SpatialStack-method}
\alias{[<-,SpatialStack,numeric,ANY,SpatialStack-method}
\alias{[[<-,SpatialStack,character,ANY,ANY-method}
\alias{[[<-,SpatialStack,logical,ANY,ANY-method}
\alias{[[<-,SpatialStack,numeric,ANY,VectorSpatialClasses-method}
\title{Replace layers of a SpatialStack object}
\usage{
\S4method{[}{SpatialStack,character,ANY,VectorSpatialClasses}(x, i, j = NULL, ...) <- value

\S4method{[}{SpatialStack,logical,ANY,VectorSpatialClasses}(x, i, j = NULL, ...) <- value

\S4method{[}{SpatialStack,numeric,ANY,VectorSpatialClasses}(x, i, j = NULL, ...) <- value

\S4method{[}{SpatialStack,character,ANY,SpatialStack}(x, i, j = NULL, ...) <- value

\S4method{[}{SpatialStack,logical,ANY,SpatialStack}(x, i, j = NULL, ...) <- value

\S4method{[}{SpatialStack,numeric,ANY,SpatialStack}(x, i, j = NULL, ...) <- value

\S4method{[[}{SpatialStack,character,ANY,ANY}(x, i, j, ...) <- value

\S4method{[[}{SpatialStack,logical,ANY,ANY}(x, i, j, ...) <- value

\S4method{[[}{SpatialStack,numeric,ANY,VectorSpatialClasses}(x, i, j = NULL, ...) <- value
}
\arguments{
\item{x}{\code{SpatialStack} object.}

\item{i}{subscript of vector-like subsetting.}

\item{j}{unused.}

\item{...}{unused..}

\item{value}{A single \code{Spatial} object.}
}
\value{
None.
}
\description{
The single and double bracket subsetting is identical in the case of SpatialStacks.
}
\examples{
data(coastlines)
spstack <- coastlines@stack[1:2]
spstack[1] <- mapedge()
}
