% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-subset.R
\docType{methods}
\name{[<-,RasterArray,ANY,ANY,logical-method}
\alias{[<-,RasterArray,ANY,ANY,logical-method}
\alias{[<-,RasterArray,ANY,ANY,RasterLayer-method}
\title{Replace layers of a RasterArray object}
\usage{
\S4method{[}{RasterArray,ANY,ANY,logical}(x, i, j, ...) <- value

\S4method{[}{RasterArray,ANY,ANY,RasterLayer}(x, i, j, ...) <- value
}
\arguments{
\item{x}{\code{RasterArray} object.}

\item{i}{subscript of the first dimension(rows) or vector-like subsetting.}

\item{j}{subscript of the second dimension (columns).}

\item{...}{subscript of additional dimensions.}

\item{value}{A \code{RasterLayer} or \code{RasterArray} object.}
}
\value{
None.
}
\description{
Single bracket \code{'['} refers to indices and names within the \code{RasterArray}. Use double brackets to replace layers based on their names (in the stack).
}
\examples{
data(dems)
# replace third element with missing value
dems[3] <- NA
# duplicate first element and make it the second too
dems[2] <-dems[1]

}
