% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-xattrib.R
\name{rownames,XArray-method}
\alias{rownames,XArray-method}
\alias{rownames<-,XArray-method}
\title{Row names of two-dimensional \code{\link{RasterArray}} or \code{\link{SpatialArray}} objects}
\usage{
\S4method{rownames}{XArray}(x)

\S4method{rownames}{XArray}(x) <- value
}
\arguments{
\item{x}{\code{\link{RasterArray}} or \code{\link{SpatialArray}} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of row names or \code{NULL}.
}
\description{
Get or set the row names of two-dimensional \code{\link{RasterArray}} or \code{\link{SpatialArray}}objects
}
\examples{
data(coasts)
rownames(coasts)
rownames(coasts) <- paste(rownames(coasts), "Ma")
}
