% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-base.R
\name{stack,VectorSpatialClasses-method}
\alias{stack,VectorSpatialClasses-method}
\alias{stack,SpatialStack-method}
\title{Stacking method for the vector Spatial* \code{\link{SpatialStack}} objects}
\usage{
\S4method{stack}{VectorSpatialClasses}(x, ...)

\S4method{stack}{SpatialStack}(x, ...)
}
\arguments{
\item{x}{\code{SpatialPoints},\code{SpatialPointsDataFrame},\code{SpatialLines},\code{SpatialLinesDataFrame},\code{SpatialPolygons},\code{SpatialPolygonsDataFrame}, object.}

\item{...}{Additional Spatial* objects.}
}
\value{
A \code{\link{RasterArray}} class object.

A \code{\link{RasterArray}} class object.
}
\description{
The function alls a \code{\link{RasterArray}}-like stacking of Spatial* objects and \code{\link{SpatialStack}}s .
}
\examples{
data(coasts)
one <- coasts[1]
two <- coasts[2]
three <- coasts[3]
# create a SpatialStack similar to a RasterStack
spStack <- stack(one, two, three)
}
