% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray.R
\docType{methods}
\name{[[<-,RasterArray,ANY,ANY-method}
\alias{[[<-,RasterArray,ANY,ANY-method}
\title{Replace RasterLayers in a RasterArray object}
\usage{
\S4method{[[}{RasterArray,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{\code{RasterArray} object.}

\item{value}{\code{character} vector.}
}
\description{
Replace RasterLayers in a RasterArray object
}
\examples{
# an example
}
