% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStack-base.R, R/SpatialArray-base.R
\name{types}
\alias{types}
\alias{types,SpatialStack-method}
\alias{types,SpatialArray-method}
\title{Return types of objects in a SpatialStack  or SpatialArray object}
\usage{
types(x)

\S4method{types}{SpatialStack}(x)

\S4method{types}{SpatialArray}(x)
}
\arguments{
\item{x}{\code{SpatialStack} object.}
}
\value{
A \code{character} class object.
}
\description{
Methods sequences that start with an NA do not yet work.
}
